+++ src/love/system.cpp
@@ -2,6 +2,7 @@
 #include "../ChaiLove.h"
 
 #include <string>
+#include "libretro.h"
 
 namespace love {
 
@@ -50,6 +51,17 @@ std::vector<int> system::getVersion() {
 
 std::string system::getVersionString() {
 	return CHAILOVE_VERSION_STRING;
+}
+
+void system::message(const std::string& msg, int frames) {
+	retro_message retroMessage;
+	retroMessage.msg = msg.c_str();
+	retroMessage.frames = frames;
+	ChaiLove::environ_cb(RETRO_ENVIRONMENT_SET_MESSAGE, &retroMessage);
+}
+
+void system::message(const std::string& msg) {
+	message(msg, 3000);
 }
 
 }  // namespace love